safari-600.7-branch/LayoutTests/accessibility/canvas-accessibilitynodeobject-expected.txt

Link  Button           ARIA button ARIA link
This test makes sure that AccessibilityNodeObjects are created for elements in a canvas subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axRenderObjects.length is axNodeObjects.length
PASS i == 0; axRenderObject.role == axNodeObject.role is true
PASS i == 1; axRenderObject.role == axNodeObject.role is true
PASS i == 2; axRenderObject.role == axNodeObject.role is true
PASS i == 3; axRenderObject.role == axNodeObject.role is true
PASS i == 4; axRenderObject.role == axNodeObject.role is true
PASS i == 5; axRenderObject.role == axNodeObject.role is true
PASS i == 6; axRenderObject.role == axNodeObject.role is true
PASS i == 7; axRenderObject.role == axNodeObject.role is true
PASS i == 8; axRenderObject.role == axNodeObject.role is true
PASS successfullyParsed is true

TEST COMPLETE
